Improve interface
